+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented here.
 
-## v1.2.0 - 2023-06-03
+## v1.2.0 - 2023-06-07
 
 ### Added
 
@@ -11,9 +11,9 @@ All notable changes to this project will be documented here.
     - For modular JavaFX applications, `JFXShaderModule.setup()` must be called before using any other of the library's
       classes.
 - Added more sample effects:
-    - [Pixelate](src/main/java/de/teragam/jfxshader/samples/effects/pixelate/Pixelate.java)
-    - [ZoomRadialBlur](src/main/java/de/teragam/jfxshader/samples/effects/zoomradialblur/ZoomRadialBlur.java)
-    - [ProxyShaderEffect](src/main/java/de/teragam/jfxshader/samples/effects/proxyshadereffect/ProxyShaderEffect.java)
+    - [Pixelate](src/main/java/de/teragam/jfxshader/samples/effects/Pixelate.java)
+    - [ZoomRadialBlur](src/main/java/de/teragam/jfxshader/samples/effects/ZoomRadialBlur.java)
+    - [ProxyShaderEffect](src/main/java/de/teragam/jfxshader/samples/effects/ProxyShaderEffect.java)
 - Added continuous rendering option to ShaderEffects.
 - EffectPeers can now invalidate their shaders to allow for a more dynamic shader pipeline.
 - Added option to create a dedicated EffectPeer for every ShaderEffect instance by setting the default `singleton`
@@ -33,7 +33,7 @@ All notable changes to this project will be documented here.
     - Applicable to any `Shape3D` by using `setMaterial`.
     - Custom parameters and textures can be set and used in the shader.
     - Pixel shaders as well as vertex shaders are supported.
-- Added [FresnelMaterial](src/main/java/de/teragam/jfxshader/samples/materials/fresnel/FresnelMaterial.java) as an
+- Added [FresnelMaterial](src/main/java/de/teragam/jfxshader/samples/materials/FresnelMaterial.java) as an
   example for a custom 3D shader material.
 
 ### Changed
@@ -47,14 +47,14 @@ All notable changes to this project will be documented here.
 
 - Shaders can now use up to 16 bound textures simultaneously.
 - Some refactoring and QoL improvements to ease the development of
-  custom [IEffectRenderer](src/main/java/de/teragam/jfxshader/IEffectRenderer.java).
+  custom [IEffectRenderer](src/main/java/de/teragam/jfxshader/effect/IEffectRenderer.java).
 - The generated texture coordinates for the shader input textures can now be accessed in
-  the [ShaderEffectPeer](src/main/java/de/teragam/jfxshader/ShaderEffectPeer.java). This allows the shaders to
+  the [ShaderEffectPeer](src/main/java/de/teragam/jfxshader/effect/ShaderEffectPeer.java). This allows the shaders to
   compensate for the varying texture coordinates caused by JavaFX's internal texture pooling.
 - The `ImagePoolPolicy.QUANTIZED` option has been added to allow better VRAM usage while ensuring that the target
   texture dimensions do not fluctuate between frames.
 - Added option to invert the mask in
-  the [BlendShapes](src/main/java/de/teragam/jfxshader/samples/effects/blendshapes/BlendShapesEffectPeer.java) example.
+  the [BlendShapes](src/main/java/de/teragam/jfxshader/samples/effects/BlendShapesEffectPeer.java) example.
 
 ### Fixed
 
@@ -77,7 +77,7 @@ All notable changes to this project will be documented here.
 - Added basic support for DPI scaling. JavaFX accomplishes this by applying a scale transform. However, this does not
   work for shaders that rely on positional information, such as the *Lighting* Effect. For now, the shaders can load the
   DPI scale as a parameter and compensate for it accordingly (see
-  the [BlendShapes](src/main/java/de/teragam/jfxshader/samples/effects/blendshapes/BlendShapesEffectPeer.java)
+  the [BlendShapes](src/main/java/de/teragam/jfxshader/samples/effects/BlendShapesEffectPeer.java)
   EffectPeer for an
   example).
