+++ src/main/java/de/teragam/jfxshader/samples/effects/Pixelate.java
@@ -1,4 +1,4 @@
-package de.teragam.jfxshader.samples.effects.pixelate;
+package de.teragam.jfxshader.samples.effects;
 
 import javafx.beans.property.DoubleProperty;
