+++ src/main/java/module-info.java
@@ -7,8 +7,6 @@ module de.teragam.jfxshader {
     exports de.teragam.jfxshader.material;
     exports de.teragam.jfxshader.exception;
     exports de.teragam.jfxshader.renderstate;
-    exports de.teragam.jfxshader.samples.effects.pixelate;
-    exports de.teragam.jfxshader.samples.effects.blendshapes;
-    exports de.teragam.jfxshader.samples.effects.zoomradialblur;
-    exports de.teragam.jfxshader.samples.materials.fresnel;
+    exports de.teragam.jfxshader.samples.effects;
+    exports de.teragam.jfxshader.samples.materials;
 }
